fix configuration script fail
google-android-build-tools-installer fail to install on my computer:

   install: impossible d'évaluer « /var/cache/google-android-build-tools-installer/android-4.4W/split-select »: Aucun fichier ou dossier de ce type

Meaning that it can find split-select.

It seem that it fail because it is not installing android-4.4W but
android-6.0

There is another bug in the Makefile: the library (libLLVM.so
libbcinfo.so libbcc.so libclang.so) are looked for in the wrong
subdirectory of android-*/, and the test looking at their possible
installation in the current system look in /usr/bin and not /usr/lib.